pyproject.toml
[tool.poetry]
name = "request-corp-sim-card-consumer"
version = "0.1.0"
description = ""
authors = ["danilkonkov <konkov@motivtelecom.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faust-streaming = "^0.8.8"
zeep = { extras = ["async"], version = "^4.1.0" }
prometheus-client = "^0.14.1"
#loguru = "^0.6.0"
loguru-ext = "^1.0.7"
pydantic = "^1.10.1"
lxml = "^4.9.1"
beautifulsoup4 = "^4.11.1"
pendulum = "^2.1.2"


[tool.poetry.dev-dependencies]


[[tool.poetry.source]]
name = "nexus"
url = "https://nexus.corp.motiv/repository/pypi-internal"
default = true
secondary = false


[[tool.poetry.source]]
name = "nexus-pypi"
url = "https://nexus.corp.motiv/repository/pypi-all/simple"
default = false
secondary = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

